refactor: 테스트 파일을 domain/repository 패키지로 분리
