+++ src/test/java/qna/repository/AnswerRepositoryTest.java
@@ -1,8 +1,11 @@
-package qna.domain;
+package qna.repository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.Answer;
+import qna.domain.AnswerRepository;
+import qna.domain.AnswerTest;
 
 import java.util.List;
