+++ src/test/java/qna/repository/DeleteHistoryRepositoryTest.java
@@ -1,8 +1,12 @@
-package qna.domain;
+package qna.repository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.DeleteHistory;
+import qna.domain.DeleteHistoryRepository;
+import qna.domain.DeleteHistoryTest;
+import qna.domain.QuestionRepository;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
