+++ src/test/java/qna/repository/QuestionRepositoryTest.java
@@ -1,8 +1,10 @@
-package qna.domain;
+package qna.repository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.Question;
+import qna.domain.QuestionRepository;
 
 import java.util.List;
