+++ src/test/java/qna/repository/UserRepositoryTest.java
@@ -1,8 +1,11 @@
-package qna.domain;
+package qna.repository;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import qna.domain.User;
+import qna.domain.UserRepository;
+import qna.domain.UserTest;
 
 import java.util.List;
